eCommerce Product Catalog 1.9.0
1.9.0

+++ theme-product_adder_support.php
@@ -13,7 +13,8 @@
 
 function al_product_adder_admin_notices_styles() {
 $template = get_option( 'template' );
-if ( ! current_theme_supports( 'ecommerce-product-catalog' ) && ! in_array( $template, array( 'twentythirteen', 'twentyeleven', 'twentytwelve', 'twentyten', 'twentyfourteen' ) ) ) {
+if ( ! current_theme_supports( 'ecommerce-product-catalog' ) && ! in_array( $template, array( 'twentythirteen', 'twentyeleven', 'twentytwelve', 'twentyten', 'twentyfourteen' ) ) && ! file_exists(get_theme_root() . '/'. get_template() . '/product-adder.php') ) {
+	$integration_type = get_integration_type();
 	if ( ! empty( $_GET['hide_al_product_adder_support_check'] ) ) {
 		update_option( 'product_adder_theme_support_check', $template );
 		return;
@@ -25,14 +26,37 @@ if ( ! current_theme_supports( 'ecommerce-product-catalog' ) && ! in_array( $tem
 }	
 add_action( 'admin_print_styles', 'al_product_adder_admin_notices_styles' );	
 
-function product_adder_theme_check_notice() { ?>
+function product_adder_theme_check_notice() { 
+if (is_integration_mode_selected() && get_integration_type() == 'simple') { ?>
 <div id="message" class="updated product-adder-message messages-connect">
 	<div class="squeezer">
-		<h4><?php _e( '<strong>Your theme does not declare eCommerce Product Catalog support</strong> &#8211; if you encounter layout issues please read our integration guide or choose a recomended theme :)', 'al-ecommerce-product-catalog' ); ?></h4>
-		<p class="submit"><a href="http://implecode.com/wordpress/product-catalog/theme-integration-guide/#cam=catalog-settings-link&key=integration-link" target="_blank" class="button-primary"><?php _e( 'Theme Integration Guide', 'al-ecommerce-product-catalog' ); ?></a> <a class="skip button" href="edit.php?post_type=al_product&page=product-settings.php&tab=product-settings&submenu=support"><?php _e( 'Plugin Support', 'al-ecommerce-product-catalog' ); ?></a><a class="skip button" href="<?php echo add_query_arg( 'hide_al_product_adder_support_check', 'true' ); ?>"><?php _e( 'Hide this notice', 'al-ecommerce-product-catalog' ); ?></a></p>
+		<h4><?php _e( '<strong>You are currently using eCommerce Product Catalog in Simple Mode</strong> &#8211; to switch to Advanced Mode you probably need Theme Integration.', 'al-ecommerce-product-catalog' ); ?></h4>
+		<p class="submit"><a href="http://implecode.com/wordpress/product-catalog/theme-integration-guide/#cam=catalog-settings-link&key=integration-link" target="_blank" class="button-primary"><?php _e( 'Theme Integration Guide', 'al-ecommerce-product-catalog' ); ?></a> <a class="skip button" href="edit.php?post_type=al_product&page=product-settings.php&tab=product-settings&submenu=support"><?php _e( 'Plugin Support', 'al-ecommerce-product-catalog' ); ?></a><a class="skip button" href="<?php echo add_query_arg( 'hide_al_product_adder_support_check', 'true' ); ?>"><?php _e( 'I know, don\'t bug me', 'al-ecommerce-product-catalog' ); ?></a></p>
 	</div>
-</div> 
-<?php }
+</div><?php
+}
+else if (is_integration_mode_selected() && get_integration_type() == 'advanced') { ?>
+<div id="message" class="updated product-adder-message messages-connect">
+	<div class="squeezer">
+		<h4><?php _e( '<strong>You are currently using eCommerce Product Catalog in Advanced Mode without the integration file</strong> &#8211; please see the guide for quick integration file creation.', 'al-ecommerce-product-catalog' ); ?></h4>
+		<p class="submit"><a href="http://implecode.com/wordpress/product-catalog/theme-integration-guide/#cam=catalog-settings-link&key=integration-link" target="_blank" class="button-primary"><?php _e( 'Theme Integration Guide', 'al-ecommerce-product-catalog' ); ?></a> <a class="skip button" href="edit.php?post_type=al_product&page=product-settings.php&tab=product-settings&submenu=support"><?php _e( 'Plugin Support', 'al-ecommerce-product-catalog' ); ?></a><a class="skip button" href="<?php echo add_query_arg( 'hide_al_product_adder_support_check', 'true' ); ?>"><?php _e( 'I know, don\'t bug me', 'al-ecommerce-product-catalog' ); ?></a></p>
+	</div>
+</div><?php
+}
+else { 
+$product_id = sample_product_id();
+$sample_product_url = get_permalink($product_id); 
+if (! $sample_product_url || get_post_status( $product_id ) != 'publish') {
+$sample_product_url = add_query_arg( 'create_sample_product_page', 'true' );
+} ?>
+<div id="message" class="error product-adder-message messages-connect">
+	<div class="squeezer">
+		<h4><?php _e( '<strong>Your theme does not declare eCommerce Product Catalog support</strong> &#8211; please proceed to sample product page where automatic layout adjustment can be done.', 'al-ecommerce-product-catalog' ); ?></h4>
+		<p class="submit"><a href="<?php echo $sample_product_url ?>" class="button-primary"><?php _e( 'Sample Product Page', 'al-ecommerce-product-catalog' ); ?></a><a href="http://implecode.com/wordpress/product-catalog/theme-integration-guide/#cam=catalog-settings-link&key=integration-link" target="_blank" class="button"><?php _e( 'Theme Integration Guide', 'al-ecommerce-product-catalog' ); ?></a> <a class="skip button" href="edit.php?post_type=al_product&page=product-settings.php&tab=product-settings&submenu=support"><?php _e( 'Plugin Support', 'al-ecommerce-product-catalog' ); ?></a><a class="skip button" href="<?php echo add_query_arg( 'hide_al_product_adder_support_check', 'true' ); ?>"><?php _e( 'Hide this notice', 'al-ecommerce-product-catalog' ); ?></a></p>
+	</div>
+</div><?php
+}
+}
 
 function implecode_product_catalog_links( $links ) {
 $links[] = '<a href="'. get_admin_url(null, 'edit.php?post_type=al_product&page=product-settings.php') .'">Settings</a>';
